Add files via upload
This is a Streamlit-based web application for detecting pneumonia from chest X-ray images using a pre-trained Convolutional Neural Network (CNN) model. The app allows users to upload an X-ray image and returns a classification — Normal or Pneumonia — along with a confidence score. The model is based on MobileNetV2 and was trained using a labeled dataset of chest X-rays from Kaggle.

This app showcases the practical application of deep learning in medical imaging and is intended for educational and demonstration purposes.